fix pausing and resuming downloads
Android was coded to reject promises when there is an existing session
for the video id provided, but we don't want that since there should
be an existing session for in progress downloads.

So instead, we find this session and pause/resume it when available.

If there isn't an existing session (which probably won't happen) then
we create a new one to execute the commands