Toggle between pause/resume based on download state
Now that we have states and listeners responding properly to the
pause and resume actions, we switch the download session pause and
resume calls to call this.onVideo(video) instead of calling the pause
and resume functions directly.

This allows us to take an action based on the download state, which is
helpful because this allows for smooth user experience even if something
becomes corrupted in the UI that would cause the wrong function to be
called as it will just toggle between them.

Now, users can pause and resume downloads in the sample app easily by
just clicking on the hourglass icon while a download is in progress